+++ src/Contracts/HydratorContract.php
@@ -4,12 +4,16 @@ declare(strict_types=1);
 
 namespace JfheinrichEu\LaravelMakeCommands\Contracts;
 
+use JfheinrichEu\LaravelMakeCommands\Dto\DataTransferObject;
+
 interface HydratorContract
 {
     /**
-     * @param class-string<DtoContract> $class
-     * @param array<string,mixed> $properties
-     * @return DtoContract
+     * fill
+     *
+     * @param  class-string<DataTransferObject> $class
+     * @param  array<string,mixed> $properties
+     * @return DataTransferObject
      */
-    public function fill(string $class, array $properties): DtoContract;
+    public function fill(string $class, array $properties): DataTransferObject;
 }