+++ tests/PackageTestCase.php
@@ -4,11 +4,14 @@ declare(strict_types=1);
 
 namespace JfheinrichEu\LaravelMakeCommands\Tests;
 
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use JfheinrichEu\LaravelMakeCommands\LaravelMakeCommandsPackageProvider;
 use Orchestra\Testbench\TestCase;
-use JfheinrichEu\LaravelMakeCommands\Providers\PackageServiceProvider;
 
 class PackageTestCase extends TestCase
 {
+    use RefreshDatabase;
+
     protected function setUp(): void
     {
         parent::setUp();
@@ -21,7 +24,12 @@ class PackageTestCase extends TestCase
     protected function getPackageProviders($app): array
     {
         return [
-            PackageServiceProvider::class,
+            LaravelMakeCommandsPackageProvider::class,
         ];
+    }
+
+    protected function defineDatabaseMigrations(): void
+    {
+        $this->loadMigrationsFrom(__DIR__ . '/database/migrations');
     }
 }